feat: implement floating CTA visibility logic and related hooks
